+++ src/attributes/pipeline.behavior.attribute.ts
@@ -1,17 +1,18 @@
 /* eslint-disable @typescript-eslint/ban-types */
 
-import {mediatorSettings} from "@/index.js";
+import { mediatorSettings } from "@/index.js";
 import type { IPipelineBehaviorClass } from "@/interfaces/ipipeline.behavior.js";
 
 /**
  * Decorate the pipelineBehavior with this attribute
- * 
+ *
  * @param value The request type
  */
-const pipelineBehavior = () => {
+const pipelineBehavior = (uniqueId?: string) => {
     return (target: Function): void => {
         mediatorSettings.dispatcher.behaviors.add({
-            behavior: target as IPipelineBehaviorClass
+            behavior: target as IPipelineBehaviorClass,
+            uniqueId,
         });
     };
 };