+++ src/interfaces/irequest.ts
@@ -9,6 +9,11 @@
 export default interface IRequest<T> {
     // eslint-disable-next-line @typescript-eslint/ban-types
     constructor: Function;
+
+    /**
+     * The request explicit name, otherwise the class name
+     */
+    readonly uniqueId?: string;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any