+++ src/models/mediator.ts
@@ -5,7 +5,7 @@ import type INotification from "@/interfaces/inotification.js";
 import type { INotificationClass } from "@/interfaces/inotification.js";
 import type INotificationHandler from "@/interfaces/inotification.handler.js";
 import type IRequest from "@/interfaces/irequest.js";
-import {mediatorSettings} from "@/index.js";
+import { mediatorSettings } from "@/index.js";
 import type IPipelineBehavior from "@/interfaces/ipipeline.behavior.js";
 
 /**
@@ -26,22 +26,23 @@ export default class Mediator implements IMediator {
      * @memberof Mediator
      */
     public async send<T>(request: IRequest<T>): Promise<T> {
-        const name = request.constructor.name;
+        const name = request.uniqueId ?? request.constructor.name;
 
         const handler = mediatorSettings.resolver.resolve<IRequestHandler<IRequest<T>, T>>(name);
-        const behaviors = mediatorSettings.dispatcher.behaviors
-            .getAll()
-            .map(p => p.behavior);
-        
+        const behaviors = mediatorSettings.dispatcher.behaviors.getAll();
+
         let currentBehaviorIndex = 0;
+
         const next = async (): Promise<T> => {
-            if(currentBehaviorIndex < behaviors.length) {
-                const behaviorClass = behaviors[currentBehaviorIndex];
-                const behavior = mediatorSettings.resolver.resolve<IPipelineBehavior>((behaviorClass as unknown as Function).name);
+            if (currentBehaviorIndex < behaviors.length) {
+                const behaviourItem = behaviors[currentBehaviorIndex];
+
+                const behavior = mediatorSettings.resolver.resolve<IPipelineBehavior>(
+                    behaviourItem.uniqueId ?? (behaviourItem.behavior as unknown as Function).name
+                );
                 currentBehaviorIndex++;
-                return await behavior.handle(request, next) as Promise<T>;
-            }
-            else {
+                return (await behavior.handle(request, next)) as Promise<T>;
+            } else {
                 return await handler.handle(request);
             }
         };
@@ -57,11 +58,17 @@ export default class Mediator implements IMediator {
      * @memberof Mediator
      */
     public async publish(message: INotification): Promise<void> {
-        const events = mediatorSettings.dispatcher.notifications.getAll(message.constructor as INotificationClass);
+        const events = mediatorSettings.dispatcher.notifications.getAll(
+            message.constructor as INotificationClass
+        );
 
-        await Promise.all(events.map(async (p) => {
-            const handler = mediatorSettings.resolver.resolve<INotificationHandler<INotification>>((p.handler as unknown as Function).name);
-            return handler.handle(message);
-        }));
+        await Promise.all(
+            events.map(async (p) => {
+                const name = p.uniqueId ?? (p.handler as unknown as Function).name;
+                const handler =
+                    mediatorSettings.resolver.resolve<INotificationHandler<INotification>>(name);
+                return handler.handle(message);
+            })
+        );
     }
 }