+++ src/interfaces/ipipeline.behavior.ts
@@ -1,11 +1,11 @@
+import type RequestData from "@/models/request-data.js";
+
 /**
  *  The pipeline behavior interface.
  *
  * @export
  * @interface PipelineBehavior
  */
-
-import type RequestData from "@/models/request-data.js";
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface, @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types, @typescript-eslint/no-unused-vars
 export default interface PipelineBehavior {