+++ src/models/notification.ts
@@ -5,7 +5,7 @@
  * @interface NotificationData
  * @template T
  */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
+
 export default class NotificationData {
 }