+++ src/models/resolver.ts
@@ -1,6 +1,10 @@
 import Resolver, { Class } from "@/interfaces/iresolver";
 
+/**
+ * The default resolver.
+ */
 export class InstantiationResolver implements Resolver {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     public add<T>(_type: Class<T>): void {
         // ignored - the instantiation resolver does not need to register anything.
     }