+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add the possibility to specify an identifier for the handler class, this is due to duplicated name when source code is minified
+- Add the possibilty to specify the identifier also for behaviour and notifications
+
 ## [1.2.1] - 2023-08-09
 
 ### Added